Major Change
Realised that the Splitting wasn't really good so stratified it according to y.
This ensured that i am getting more uniform splits wrt the data.

Secondly, I had only used accuracy as a metric which would be wrong. Now the code includes a classification report and also confusion matrix (although only for KNN for now)

Looking foreward to learning more!